cocoa_common: update the screen info before entering fullscreen
When the user was not specifing any -xineramascreen, the mplayer2
window was entering fullscreen in the screen where occurred the video
output initialization. With this fix the window chooses which screen
to go fullscreen in depending on it's current position.

This doesn't change the fact that the behaviour can be overriden using
the -xineramascreen option by passing the desired screen id to go
fullscreen in.