vo_sharedbuffer: add this video output
This video output is based on -vo corevideo:shared_buffer and manages a
shared buffer and a Cocoa distributed object to communicate with GUIs.

Extracting this code to a different video output makes sense because the
gui-less corevideo was following a completely different code path from
the shared buffer code. Even so, some parts were shared between the two
making it difficult to maintain and evolve the main vo_corevideo code.