vo_sharedbuffer: add this video output
This video output is based on -vo corevideo:shared_buffer and manages a
shared buffer and a Cocoa distributed object to communicate with GUIs.

Extract this code to a different video output to get rid of harmful code
coupling, performance inefficiencies (i.e.: useless image memory copies)
and ugly code (big if-else conditionals).