vo_gl3: request OpenGL 3.0 context, and add hack to deal with GL 3.2 on OSX
Until now, a 3.2 context was requested, even though the code requires
only OpenGL 3.0. Until now, 3.2 was requested, because OSX doesn't
support 3.0; you always get a 3.2 context. The problem is that shaders
must declare the GLSL version they are using, making it hard to make a
shader work on both 3.0 and 3.2 (unless the 3.2 context has some
backward compatibility, which OSX doesn't have). So a shader declared
with declared with version 3.0 (that is, using a "#version 130"
directive) won't work on OSX. Add a stupid hack that checks whether we
actually got a 3.2 context (i.e. probably on OSX), and use a higher
shader version if it is.