cocoa_common: fix alt tabbing to current and from different workspace
Fix alt tabbing when tabbing to a window in the same workspace: The window
stayed on top because of a missing call to orderBack:.

Fix alt tabbing when tabbing from a window in a different workspace: The window
didn't get activated. Turns out that you must call makeKeyAndOrderFront:
before setLevel: or setPresentationOptions: or the window will not
properly ask for focus.