cocoa_common: fix problems with alt-tab window changes
Fix alt tabbing to another window in the same workspace. The player
window stayed on top because of a missing call to orderBack:.

Fix alt tabbing to the player window from a different workspace. The
window didn't get activated. Turns out that you must call
makeKeyAndOrderFront: before setLevel: or setPresentationOptions: or
the window will not properly ask for focus.