m_config: support flatten and merge flags for suboptions
M_OPT_FLATTEN allows adding top-level options with the suboption
mechanism. The point of this is that, even though these options are
top-level options, they don't need to be added directly to a top-
level option array (such as mplayer_opts[]). Instead, the suboption
can be defined in a separate source file. Only a suboption declaration
is needed to add these options. The intention is to promote
modularization. In confunction with sub-struct support, to remove the
need for one global monolithic option struct, namely MPOpts.

M_OPT_MERGE is similar to M_OPT_FLATTEN, but doesn't add the name of
the suboptions entry as prefix.

Given you have a suboption declaration "prefix" and the suboption
"subopt", you can pass them as follows on the command line:
  normal:           -prefix subopt=value
  M_OPT_FLATTEN:    -prefix-subopt value
  M_OPT_MERGE:      -subopt value