cfg-mplayer, screenshot: parse image options separately
This means we don't have to duplicate all image writer options. It
makes it possible to add new options to the image writer code, and
making them available as new --screenshot-* options, without having
to edit the screenshot code (and as soon as it is added, the vo_image
code).