OSX: configure: fix check for additional cflags
Currently, the test to add the extra cflags is done against $_cc and
not $cc_vendor. This works ok until the system's cc is gcc.

Recent versions of Xcode changed the default compiler (cc) to clang.
This causes a lot of spam from clang, who complains about the unknown
flags, when mplayer is configured with no --cc option.