cocoa_common: conditionally hide menubar and dock when going fs
The current implementation was hiding the dock and menubar unconditionally
when going fullscreen. This means they were hidden even if they weren't on
the screen mplayer2 was going fullscreen to, resulting in poor user experience.

Change the fullscreen function in the cocoa backend to check that mplayer2
is in the same screen as the menubar/dock before hiding them.