cocoa_common: add HiDPI/retina support
When using a HiDPI screen – for performance and backwards compatibility
reasons – AppKit requests an OpenGL surface assuming a pixel number that
equals the user points number.

After the image is rendered to this smaller surface, the size of the
rendered view is upscaled so that it's dimensions are comparable across
screens of different DPIs. The applied scaling is not that good and makes
the video/subtitles blurry: this is not acceptable for a video player.

This commit allows AppKit to request a high resolution OpenGL surface
to back the mplayer2 OpenGL view. It also reports back to the player the
correct pixel size of the target window by converting user points to actual
pixels.

All the system version checks are done at runtime so that the feature is
available on OSX 10.7 even with a binary compiled with older SDKs.